refactor: remove extension monitor
Remove extension monitor, removing an unnecessary layer of abstraction.
