+++ Sources/TunnelMonitorKit/TunnelMonitorKit.swift
@@ -1,6 +1,6 @@
 //
-//  File 2.swift
-//  
+//  TunnelMonitorKit.swift
+//  TunnelMonitorKit
 //
 //  Created by Chris J on 23/03/2022.
 //  Copyright © 2022 Chris Janusiewicz. Distributed under the MIT License.
